Move server impls into submodule
